ui changes and use spring security

+++ src/main/java/com/example/todowebapp/controller/TodoController.java
@@ -1,13 +1,22 @@
 package com.example.todowebapp.controller;
 
+import java.text.SimpleDateFormat;
 import java.util.Date;
+
 
 import javax.validation.Valid;
 
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.propertyeditors.CustomDateEditor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.WebDataBinder;
+import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -25,18 +34,36 @@ public class TodoController {
 	@Autowired
 	TodoService service;
 	
+	@InitBinder
+	public void initBinder(WebDataBinder binder) {
+		// Date - dd/MM/yyyy
+		SimpleDateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy");
+		binder.registerCustomEditor(Date.class, new CustomDateEditor(
+				dateFormat, false));
+	}
+	
 	@RequestMapping(value="/list-todos",method = RequestMethod.GET)
 	public String showTodos(ModelMap model) {
 		System.out.print(model);
-		String name = (String) model.get("name");
+		String name = getLoggedInUserName(model);
 		//name = "in28Minutes";
         model.put("todos", service.retrieveTodos(name));
         return "list-todos";	
 	}
 	
+	private String getLoggedInUserName(ModelMap model) {
+		Object principal = SecurityContextHolder.getContext()
+				.getAuthentication().getPrincipal();
+
+		if (principal instanceof UserDetails)
+			return ((UserDetails) principal).getUsername();
+
+		return principal.toString();
+	}
+	
 	@RequestMapping(value="/add-todo",method = RequestMethod.GET)
 	public String showAddTodoPage(ModelMap model) {
-		model.addAttribute("todo", new Todo(0, (String) model.get("name"), "Default Desc",
+		model.addAttribute("todo", new Todo(0, getLoggedInUserName(model), "Default Desc",
 				new Date(), false));
 		return "todo";
 	}
@@ -44,17 +71,18 @@ public class TodoController {
 	
 	@RequestMapping(value="/add-todo",method= RequestMethod.POST)
 	public String addTodo(ModelMap model,@Valid @ModelAttribute("todo") Todo todo,BindingResult result) {
-		System.out.print("hello   ");
 		if(result.hasErrors()){
-			//System.out.print("helloFCS");
 			return "todo";
 		}
-	   service.addTodo((String)model.get("name"),todo.getDesc(),new Date(), false);	
+	   service.addTodo(getLoggedInUserName(model),todo.getDesc(),new Date(), false);	
+	   model.clear();
 	   return "redirect:/list-todos";
 	}
 	
 	@RequestMapping(value="/delete-todo",method= RequestMethod.GET)
     public String deleteTodo(@RequestParam int id) {
+		
+		if(id == 1) throw new RuntimeException("Something went wrong");
     	service.deleteTodo(id);
     	return "redirect:/list-todos";
     }
@@ -72,11 +100,10 @@ public class TodoController {
 			return "todo";
 		}
 		
-		todo.setUser((String) model.get("name"));
+		todo.setUser(getLoggedInUserName(model));
 		todo.setTargetDate(new Date());
 		service.updateTodo(todo);
-		System.out.print("hello   ");
-
+		model.clear();
 		return "redirect:/list-todos";
 	}
 }
